+++ kinesis/src/com/splunk/modinput/kinesis/JSONOnlyMessageHandler.java
@@ -9,7 +9,7 @@ import com.splunk.modinput.kinesis.KinesisModularInput.MessageReceiver;
 public class JSONOnlyMessageHandler extends AbstractMessageHandler {
 
 	@Override
-	public void handleMessage(String record, String seqNumber,
+	public void handleMessage(String record, byte [] rawBytes,String seqNumber,
 			String partitionKey, MessageReceiver context) throws Exception {
 
 		String text = stripNewlines(record);